beginning stream read refactor
